move challenge view to containers
