+++ stem-explorer-ng/src/app/app.module.ts
@@ -23,7 +23,7 @@ import { SplashScreenComponent } from './components/splash-screen/splash-screen.
 import { ZXingScannerModule } from '@zxing/ngx-scanner';
 import { FilterButtonsComponent } from './components/filter-buttons/filter-buttons.component';
 import { FilterPipe } from './shared/pipes/filter.pipe';
-import { ChallengeViewComponent } from './components/challenge-view/challenge-view.component';
+import { ChallengeViewComponent } from './containers/challenge-view/challenge-view.component';
 import { ToolbarComponent } from './components/toolbar/toolbar.component';
 import { CameraComponent } from './containers/camera/camera.component';
 import { CameraButtonComponent } from './containers/camera-button/camera-button.component';
