+++ stem-explorer-ng/src/app/containers/challenge-view/challenge-view.component.html
